requirements.txt
pandas
numpy
psycopg2-binary
schedule
logging